requirements.txt
streamlit>=1.33.0
anthropic>=0.25.0
python-dotenv>=1.0.0
